Add better limit handling to system: when a limit is triggered, return to the linear-zero position.
Add an initialize function for startup: use the limit switches to find the track extent and then go to zero.

With the above two additions, we should be able to assume that new tests will always start at the zero position.

Also add an initialize file which just calls sys.initialize to find the zero position. This should only be run on power-up (eventually, we should configure the Pi to run this on OS startup).